improve server-side caching (#3)
The caching process has been completely rewritten and is no longer horrible to look at (thank god).

Additionally, the caching methodology in `main_controller` has been changed. Now it will run the templating engine only after the cached item is reloaded, and then it will store the output over the file's raw data. And all the whitespace is stripped from all of the parts of the site when stored in cache, which helps offset the increased component size of the cache.

Overall, the site will use a little bit more cache, but no longer has to run the templating engine on the same input every time, so it's probably a bit easier/faster.